+++ compiler_opt/rl/regalloc/gin_configs/ppo_nn_agent.gin
@@ -3,7 +3,7 @@ import compiler_opt.rl.constant
 import compiler_opt.rl.constant_value_network
 import compiler_opt.rl.gin_external_configurables
 import compiler_opt.rl.regalloc.config
-import compiler_opt.rl.regalloc_network
+import compiler_opt.rl.regalloc.regalloc_network
 import tf_agents.agents.ppo.ppo_agent
 import tf_agents.networks.actor_distribution_network
